merge from Frontend/feature

+++ backend/src/orders/orders.controller.ts
@@ -7,7 +7,6 @@ import {
   Patch,
   Param,
   Delete,
-  ForbiddenException,
   HttpCode,
   Query,
 } from '@nestjs/common';
@@ -50,8 +49,9 @@ export class OrdersController {
   filterOrders(
     @Query('startDate') startDate?: string,
     @Query('endDate') endDate?: string,
+    @Query('orderNumber') orderNumber?: string,
     @Query('customer') customer?: string,
-    @Query('userId') userId?: number,
+    @Query('user') user?: string,
     @Query('sortBy') sortBy?: string,
     @Query('sortOrder') sortOrder?: 'asc' | 'desc',
     @Query('page') page?: number,
@@ -59,8 +59,9 @@ export class OrdersController {
     return this.ordersService.filter({
       startDate: startDate ? new Date(startDate) : undefined,
       endDate: endDate ? new Date(endDate) : undefined,
+      orderNumber: orderNumber ? orderNumber : undefined,
       customer: customer ? customer : undefined,
-      userId: userId ? Number(userId) : undefined,
+      user: user ? user : undefined,
       sortBy,
       sortOrder,
       page: page ? Number(page) : undefined,
@@ -87,8 +88,9 @@ export class OrdersController {
   @Delete(':id')
   @ApiOkResponse({ type: OrderEntity })
   delete(@Param('id') id: string,
-    @GetCurrentUser('role') role: string) {
-    return this.ordersService.delete({ id }, role);
+    @GetCurrentUser('role') role: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.delete({ id }, role, userId);
   }
 
   @Post('/detail/:orderDetailId/print')
@@ -103,8 +105,10 @@ export class OrdersController {
   @Post('/detail/:orderDetailId/cancel-print')
   @Roles(Role.Admin, Role.Operator)
   @HttpCode(200)
-  cancelMarkPrinted(@Param('orderDetailId') id: string,) {
-    return this.ordersService.cancelStatus({ type: CancelType.PRINT, orderDetailId: id });
+  cancelMarkPrinted(
+    @Param('orderDetailId') id: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.cancelStatus({ type: CancelType.PRINT, orderDetailId: id, userId });
   }
 
   @Post('/:id/pay')
@@ -118,8 +122,9 @@ export class OrdersController {
   @Post('/:id/cancel-pay')
   @Roles(Role.Admin, Role.Administrasi)
   @HttpCode(200)
-  cancelMarkPay(@Param('id') id: string) {
-    return this.ordersService.cancelStatus({ type: CancelType.PAY, orderId: id });
+  cancelMarkPay(@Param('id') id: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.cancelStatus({ type: CancelType.PAY, orderId: id, userId });
   }
 
   @Post('/:id/taken')
@@ -133,7 +138,8 @@ export class OrdersController {
   @Post('/:id/cancel-taken')
   @Roles(Role.Admin, Role.Administrasi)
   @HttpCode(200)
-  cancelMarkTaken(@Param('id') id: string) {
-    return this.ordersService.cancelStatus({ type: CancelType.TAKEN, orderId: id });
+  cancelMarkTaken(@Param('id') id: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.cancelStatus({ type: CancelType.TAKEN, orderId: id, userId });
   }
 }
